Fixed "Maximum number of directories is 50"
I ran into an issue when creating a folder on mathcha. It seems that we are restricted on the number of folders we can create. My solution of having one link per folder is no longer possible, I have to connect with my login and password to get the file links individually.